requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic[email]==2.9.2
python-multipart==0.0.12
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
email-validator==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
supabase==2.3.0
httpx==0.26.0
Pillow==10.1.0
